add RxFingerprintCompat
